requirements.txt
asn1crypto==0.23.0
attrs==17.3.0
Automat==0.6.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.1.3
cssselect==1.0.1
enum34==1.1.6
Flask==0.12.2
future==0.16.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
parsel==1.2.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.4.0
python-telegram-bot==8.1.1
queuelib==1.4.2
requests==2.18.4
Scrapy==1.4.0
scrapyd==1.2.0
scrapyd-client==1.1.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
urllib3==1.22
w3lib==1.18.0
Werkzeug==0.12.2
zope.interface==4.4.3
